Move organizers data into top-level _data.json.
This commit also updates the link to these data in the about page.

The top level _data.json is a more logical place for these important
data, which might be re-used in many spots in the website. These data
structures may in the future become larger, as we need to access more
information about these individuals.

These people may be referenced as authors in blog posts, hosts of
events, etc.